requirements.txt
Django==2.2.7
django-channels==0.7.0